presentation-translator: introduce define-selection-translator macro
Selection translators are simplified presentation-translators. They
are always bound to a gesture :select, tester is always definitive and
does not depend on arguments (context-type frame event window x y)
which can't appear on the arglist. That means that selection
translators may be used in a processing context which doesn't involve
frame and a pointer (i.e in selection/clipboard protocol).